Base class file rename
