+++ tests/database/postgres/test_banned_channels.py
@@ -1,5 +1,5 @@
 from tests.database.postgres.test_database import TestPostgres
-from tests.database.tests.test_banned_channels import TestBannedChannels
+from tests.database.tests.banned_channels import TestBannedChannels
 
 
 class TestPostgresBannedChannels(TestBannedChannels, TestPostgres):
