+++ tests/database/postgres/test_game_abbreviations.py
@@ -1,5 +1,5 @@
 from tests.database.postgres.test_database import TestPostgres
-from tests.database.tests.test_game_abbreviations import TestGameAbbreviation
+from tests.database.tests.game_abbreviations import TestGameAbbreviation
 
 
 class TestPostgresGameAbbreviation(TestGameAbbreviation, TestPostgres):
