+++ tests/database/postgres/test_timezone.py
@@ -1,5 +1,5 @@
 from tests.database.postgres.test_database import TestPostgres
-from tests.database.tests.test_timezone import TestTimezone
+from tests.database.tests.timezone import TestTimezone
 from source.database import DatabaseTimeZone
