+++ tests/database/sqlite/test_auto_join.py
@@ -1,5 +1,5 @@
 from tests.database.sqlite.test_database import TestSqlite
-from tests.database.tests.test_auto_join import TestAutoJoin
+from tests.database.tests.auto_join import TestAutoJoin
 
 
 class TestSqliteAutoJoin(TestAutoJoin, TestSqlite):
