+++ tests/database/tests/auto_join.py
