+++ tests/database/tests/chat_properties.py
